implemented double jump vfx
Tried to implement my character vfx, was only able to succesfully put in the jump one. And even with that I had to change it from a spawn rate to a spawn burst so that it didn't follow the player, so i might have to make a texture for it to make it look better. Am going to need help implementing stuff from the programmers, they probably know how to do it