pushing new animations, knight, etc...
Won't really show for the showcase tomorrow/today for shbeeb, but the cates now alternate between their different talk animations when interacting with them. In the levels they should now appear as scared, and the ones that're for the time trials aren't affected by it. (from my brief testing. I basically did a check for if they were in the hub level or not, then had them do the scared animation.)
Most of the knight's new animations are in (gonna try to do death and damaged either later this week or next week, the list continues to grow), I didn't touch anything with his blueprint since I'm unsure as to where things may need to be readjusted for it, leaving that to code unless they need any help. Those new animations are under boss_3, animations, then their own folder inside of that.
And riley's victory animation for level completion is in as a sequence. Not in her animation blueprint since I think that would work as a montage? Not sure how that's planned on being implemented, but it seems more like a sequence/cinematic thing, so I'd imagine it would be best as an animation montage? if that's the case, that's a one-button thing and doesn't need anything from my end, anyone can do that and I won't mind.